Code cleanup: dependencies, API surface, Swagger accuracy

+++ src/PaymentGateway.Application/Models/ProcessPaymentRequest.cs
@@ -6,5 +6,5 @@ public record ProcessPaymentRequest(
     int ExpiryMonth,
     int ExpiryYear,
     string Currency,
-    int Amount,
+    long Amount,
     string Cvv);
